requirements.txt
# requirements.txt
streamlit
google-cloud-firestore
requests
google-generativeai
python-dotenv